Optimize Actor Loss in Causal RL Algorithm by Switching from Mean to Sum Reduction
modified the reduction method in the computation of the actor loss from a mean to a sum. This change addresses an issue where the mean reduction of advantage * invertible_action_error sometimes resulted in a minimal loss, especially when the advantage values varied in sign. This minimal loss was insufficient in motivating the reinforcement of agent learning actions.

Extensive experiments have shown that using a sum reduction not only enhances the learning speed but also maintains good generalization across various scenarios. This update is applied exclusively to the actor loss component, ensuring a more robust and efficient learning process in the RL algorithm. Additionally, this modification has enabled a reduction in the minimum required batch size for learning the same environments.