Change value loss function from MSE to Smooth L1 (Huber Loss)
The value loss function used in our model from Mean Squared Error (MSE) to Smooth L1 Loss, also known as Huber Loss. The primary motivation for this change is to reduce the sensitivity of our loss function to outliers. Huber Loss combines the best properties of L1 and L2 loss, being less sensitive to outliers than MSE. This should enhance the stability and robustness of our model, especially in scenarios with noisy data or significant variations in value estimates.