Refined Value Normalization in Causal RL
Implemented an improved normalization strategy for both actual returns(expected_value) and value estimates in the Causal RL framework. By updating the normalization based on the actual returns (value) and applying this normalization to the value estimates, we ensure consistency in scale and reduce variance across the learning process. This approach not only aligns the scale of returns and value estimates but also enhances learning stability by managing discrepancies and fluctuations in reward magnitude and distribution effectively.