Causal RL v1.1.10, Update Benchmarks: https://wandb.ai/causal-rl/causal-rl/
In this release of Causal RL, we've transitioned the loss reduction strategy to "cross" reduction from the previous "batch" reduction. This change allows for a more optimization approach by concatenating reductions across both batch and sequence dimensions, aiming to enhance reward optimization per step and per trajectory.

However, initial observations indicated that the "cross" reduction method might halve the gradient during backpropagation. To address this potential gradient diminishment, we've integrated the GradScaler class with a scale_factor set to 1. This addition ensures the preservation of gradient magnitude, particularly for the batch dimension, mitigating the risk of losing gradient information during optimization.

Furthermore, while the running absolute mean has been the default choice for advantage normalization, we're revisiting the use of running_mean_std as an alternative. To accommodate the dynamic nature of running normalizers (mean_std, abs_mean, square_mean), we've adjusted the update decay rate to 1e-3. This tweak aims to balance the responsiveness of the normalizer to recent observations while maintaining stability over time.